Add files via upload
This Python file contains an introductory Jupyter notebook on IBM Watson Studio by Muhammed Dayo Alabi, a data scientist. The notebook includes a code example that determines whether a person is an adult and eligible to drive, as well as various Markdown elements such as headings, lists, and tables. Use this notebook as a starting point for your own data science journey!