plotting_histograms
This code demonstrates how to plot a histogram using the Boston real estate dataset from sklearn. The dataset is loaded using the `load_boston()` function, and the target variable and data are assigned to `y_axis` and `x_axis`, respectively. The `matplotlib` library is used to create a histogram of the target variable `y_axis`. The `style` function is used to specify the style of the plot, and the `plt.figure()` function is used to set the size of the plot. The resulting histogram shows the distribution of the target variable `y_axis` in the Boston real estate dataset.