exploratory_data_analysis
Problem Statement:
Analyze the FAA using pandas to do the following
- view: aircrafte manufacturer, state, aircraft model, test information, flight phase, event description type, fatal flag
- clean the dataset and replace the NaN with No
- find the different types of aircrafts & their occurrences in the dataset
- remove all observation where aircraft names are not available
- display the observation with fatal flag as yes.
.
.
This code reads in a FAA (Federal Aviation Administration) dataset in csv format using the pandas library. It selects specific columns from the dataset and removes any empty cells in the selected column. It then groups the data by aircraft make and fatal flag, and prints the size of each group. Finally, it extracts the group with fatal accidents and prints the data. The code also includes various print statements to display the shape, columns, data types, and first few rows of the dataframes. Additionally, it replaces empty cells in the 'FATAL_FLAG' column with 'No' and drops empty cells in the 'ACFT_MAKE_NAME' column. Finally, it exports the 'fatal accidents' group to a new csv file.