exploratory_data_analysis
This code reads in a CSV file named "FDNY.csv" and loads it into a pandas DataFrame. The data is then explored using various pandas functions. The `describe` function is used to generate summary statistics of the DataFrame. The `head` function is used to view the first 5 rows of the data.

Next, a new DataFrame is created by skipping the first row of the CSV file, which contains the column names. This DataFrame is also explored using `describe`, `head`, and `count` functions to get an overview of the data.

The `groupby` function is used to group the data by the "Borough" column, and the `size` function is used to count the number of records in each group. The data for Manhattan borough is extracted using `get_group` function and stored in a new DataFrame named `file_info_manhattan`.

Overall, this code is an example of how to load and explore data from a CSV file using pandas. It can be useful for data analysts and data scientists who are interested in working with data stored in CSV format.