WebScrapping using BeautifulSoup
This Python code utilizes pandas, requests, and BeautifulSoup to scrape data from a Netflix stock webpage and store it in a DataFrame. The code accesses the webpage's HTML, extracts data from the table, and assigns it to dictionary keys. The dictionary keys are then used to create a DataFrame, which is printed to the console.